Updated midrise apartment schedules
The midrise apartment and office heating and cooling thermostat setpoint
schedules didn't match the ref building IDF. I fixed them by hand in the
master schedules file. I also identified that corridors  have their on
schedules, which was a unit heater. I added this and updated the
spreadsheet and JSON file to reflect this addition.