First commit of the Performance Rating Method Testing by LBNL
In this test, we set up prototype models (3 for the moment) that are variation of the DOE reference models and run tests on the model in output of the measure to check for errors or missing features.
The current commit contains 3 models that check parts of the measure for: Run_01 - Envelope, Run_12 - Lighting, Run_18 - HVAC. The ruby script checks if some of the tricky changes to be done on the baseline model are done properly.