Fix Off Cycle Loss Coefficient to Ambient Temperature
The proposed changes update the off cycle loss coefficient to ambient temperature for all prototypes (the ones that required an update). A test for service hot water was also added. As of now, it only test for the proposed changes but it will be updated at a later time to test for other updates.