Ignore Exception b to Section 9.4.1.4 to 90.1-2010
Exception b to Section 9.4.1.4 to 90.1-2010 should be ignored when generating the DOE prototypes. There is no code  requirement in 90.1-2010 for visible transmittance hence, the sidelighted effective aperture area for prototypes is not evaluated. The proposed change uses the `@instvarbuilding` variable, that is only defined when creating a prototype, to skip the sidelighted effective aperture exception test. This should not impact the baseline model creation for performance rating methods as the variable isn't defined and as a result `@instvarbuilding_type.nil?` will always be `true`.