make all unitary system object fans OnOff
Constant volume fans and cycling fans will now both be the OnOff fan object type, with the difference being the SupplyAirFanOperatingModeSchedule.
Set the default for SupplyAirFanOperatingModeSchedule to be always off to follow load.  There is code later in applying hvac efficiency standard to set this schedule to follow the zone occupancy schedule if this system provides ventilation.