rubocop edits
This is not a nice commit.  It touches lots of files and will make some code blocks a pain to debug.  But it will improve formatting in the long term.

There are more errors/warnings remaining.

I blocked out the NECB standards and increased the ABC and block length thresholds to reduce some warnings which won't be addressed soon.