Nrcan 320 (#1103)
Added CLI support to standards project.. Added support to produce btap_data.json without costing. This is required for our partners that will not have access to the costing aspects of btap and will use the public version of our work. Issue with PHIUS reporting will need to be examined by SG.