Fix regressions with pressure driven water systems
The standards method pump_standard_minimum_motor_efficiency_and_size originally was written as:
`return [1.0, 0] if motor_bhp == 0.0`
For service water heating systems that are pressure driven, is there is no pump, and the pressure is set to some tiny but no-zero value. The zero isn't catching these small values because motor bhp is still greater than zero (even if 1^-10), and so isn't triggering the if statement to return 1.0 as it should.
`return [1.0, 0] if motor_bhp < 0.0001` fixes this issue, but it changes the results. The fix is to update the regression tests with the correct value for motor efficiency in this case (1.0) and update the performance results too.
An interesting case where rubocop caught an error. Floating point comparisons with == 0.0 don't work as intended!