Handle zero airflow case for damper effectiveness
The air_loop_hvac_adjust_minimum_vav_damper_positions method expects that zones have outdoor air and design air flow rate. If the space does not have a DesignSpecificationOutdoorAir object, or it is zero, then the calculation is unnecessary. If either the zone CoolingDesignAirFlowRate or HeatingDesignAirFlowRate are zero, then the zone discharge air fraction (z_d) could have a divide by zero and result in NaN. This crashes the code when trying to calculate the minimum zone ventilation effectiveness (e_vz).

Add informative error messages that the zone CoolingDesignAirFlowRate or HeatingDesignAirFlowRate are zero, and default the zone discharge air fraction (z_d) to zero if either the zone minimum discharge airflow rate (v_dz) is zero, or the zone outdoor airflow rate is zero (v_oz).