update regression and performance results
The data center WSHPs in large offices (bottom and mid) were too large and therefore using default efficiencies. Now they still get assigned a value, equal to the maximum equipment size available.

They are not getting treated as a computer room air conditioner, which is a separate issue and may be addressed in another branch.