make supermarket schedules constant
Set constant values for heating and cooling schedules. The original values were from end use data, which was an average. It appears some stores had setbacks and most didn't, and the results is slight (1C) setbacks overnight. That causes issues in parametric schedules and sawtooth schedules. This commit changes the values to a constant average temperature.