add defrost and dripdown schedule variation
add defrost and dripdown schedule variation to refrigerated cases and walkins. Iterate the start hour over the cases and walkins added to the model. Add fields for defrosts per day and duration to the typical cases and walkins data. Assumptions for these values are placeholder values for now, based on prior values and a quick internet search.

The purpose for adding variation is so that all the schedules aren't synced and cause a spiky profile in demand.